+++ cmd/cli/root_command.go
@@ -1,7 +1,7 @@
 package cli
 
 import (
-	"github.com/cduggn/ccexplorer/internal"
+	"github.com/cduggn/ccexplorer/internal/config"
 	"github.com/common-nighthawk/go-figure"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
@@ -17,7 +17,7 @@ var (
 )
 
 func RootCommand() *cobra.Command {
-	internal.LoadConfigFunc(".")()
+	config.LoadConfigFunc(".")()
 	Initialize()
 	return rootCmd
 }
