+++ cmd/cli/validator.go
