--- internal/config.go
@@ -1,13 +1,0 @@
-package internal
-
-import "github.com/spf13/viper"
-
-var LoadConfigFunc = func(path string) func() {
-	return func() {
-		LoadConfig(path)
-	}
-}
-
-func LoadConfig(path string) {
-	viper.AutomaticEnv()
-}
